+++ proteins.config.js
@@ -55,11 +55,15 @@ module.exports = {
     // How many seeds will we insert/delete per transaction
     seedsPerBatch: 20,
 
+    // How many wildcards will we insert/delete per transaction
+    wildcardsPerBatch: 50,
+
     // Check the ./datasets folder to see which dataset sizes are available.
     files: {
       //Remember to always match the protein sizes with the seed sizes.
       proteinsInsertMany: "datasets/proteins/protein_structs_100.txt",
       seedsInsertMany: "datasets/seeds/seed_3_structs_100.txt",
+      wildcardsInsertMany: "datasets/wildcards/wildcard_3_structs_100.txt",
 
       //This allows us to add extra positions to (known) seeds! Useful if you need to add the seed's array in parts!
       seedsInsertManyPositions: "datasets/seeds/",